move ci-info to server/
