+++ packages/next/src/telemetry/anonymous-meta.ts
@@ -2,7 +2,7 @@ import isDockerFunction from 'next/dist/compiled/is-docker'
 import isWslBoolean from 'next/dist/compiled/is-wsl'
 import os from 'os'
 
-import * as ciEnvironment from './ci-info'
+import * as ciEnvironment from '../server/ci-info'
 
 type AnonymousMeta = {
   systemPlatform: NodeJS.Platform
